refactor (LottoTest) : 관련 테스트 클래스들의 패키지 위치를 조정한다
