+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.domain.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
